Take exif orientation tag into account
Recent versions of gthumb (and probably other software as well)
have decided that the "rotate" button would not rotate the image
but only set an exif tag which tells image viewers to rotate the pixel
data before displaying the image. As a result, when loading such a
rotated image in imagination, the image would be displayed unrotated.

This commit fixes this. When adding a new image in a slideshow, the exif
tag is read and the raw image is rotated automatically (as if the user
had clicked on the corresponding button).

As a result, this change is backward compatible. Existing .img files
containing images with an Orientation exif tag will render just as
before, because the tag is only read when adding an image to a slideshow
and not when loading an existing slideshow.